Update UI.xml
Further minor corrections have been made. 
Question, line 1003 and 1057 look different, is the code used correct?
All in all a very good translation. It only lacked commas and here and there plural where it had to be singular.
I didnt find the thing I was originally looking for. The definition of the times in german, where can I find those?